Update tests #267
